// src/types/allUsers.types.ts
export interface allUsers {
    id: string;
    isAdmin: boolean;
    createdAt: string;
    agente: string;
    imgUrl: string;
    nombre: string;
    telefono: string;
    direccion: string;
    latitud: number;
    longitud: number;
    documento: number;
    email: string;
    razonSocial: string;
    codigoPostal: string;
    domicilioInstal: string;
    localidadInstal: string;
    telefonoInstal: string;
    emailInstal: string;
    observaciones: string;
    senalConexion: string;
  }
  